pusher_wishlist

+++ database/migrations/2024_10_09_052141_create_banners_table.php
@@ -9,7 +9,7 @@ class CreateBannersTable extends Migration
     public function up()
     {
         Schema::create('banners', function (Blueprint $table) {
-            $table->unsignedBigInteger('banner_id')->primary();
+            $table->increments('banner_id')->primary();
             $table->string('banner_subtitle');
             $table->string('banner_title');
             $table->text('banner_description');
